Extract ReCaptcha to separate service
